requirements.txt
langchain==0.2.5
langchain-openai
python-dotenv